+++ ThinkingHome.Plugins.Scripts.WebUi/Resources/list.js
@@ -61,7 +61,7 @@ var Section = lib.common.AppSection.extend({
         this.view.showChildView('list', listView);
     },
     displayError: function (title, error) {
-        this.application.showError(title, error.message);
+        this.application.showErrorPage(title, error.message);
     },
 
     addScript: function () {